[Reporting] refacto slices and add button in Semaphore pop-up to go to reportings list
