+++ frontend/src/domain/use_cases/map/updateMapInteractionListeners.ts
@@ -1,6 +1,7 @@
+import { attachMissionToReportingSliceActions } from '@features/Reportings/components/ReportingForm/AttachMission/slice'
+
 import { closeAreaOverlay } from './closeAreaOverlay'
 import { attachReportingToMissionSliceActions } from '../../../features/missions/MissionForm/AttachReporting/slice'
-import { attachMissionToReportingSliceActions } from '../../../features/Reportings/slice'
 import { resetInteraction } from '../../shared_slices/Draw'
 import { setDisplayedItems, resetLayoutToDefault } from '../../shared_slices/Global'
