+++ frontend/src/domain/use_cases/reporting/attachMission.ts
@@ -1,5 +1,6 @@
+import { attachMissionToReportingSliceActions } from '@features/Reportings/components/ReportingForm/AttachMission/slice'
+
 import { missionsAPI } from '../../../api/missionsAPI'
-import { attachMissionToReportingSliceActions } from '../../../features/Reportings/slice'
 import { setToast } from '../../shared_slices/Global'
 
 export const attachMission = (id: number) => async (dispatch, getState) => {
