+++ frontend/src/domain/use_cases/reporting/createMissionFromReporting.ts
@@ -1,10 +1,10 @@
+import { reportingActions } from '@features/Reportings/slice'
 import omit from 'lodash/omit'
 
 import { reportingsAPI } from '../../../api/reportingsAPI'
 import { mainWindowActions } from '../../../features/MainWindow/slice'
 import { isNewReporting } from '../../../features/Reportings/utils'
 import { ReportingContext, setReportingFormVisibility, setToast, VisibilityState } from '../../shared_slices/Global'
-import { reportingActions } from '../../shared_slices/reporting'
 import { MapInteractionListenerEnum, updateMapInteractionListeners } from '../map/updateMapInteractionListeners'
 import { addMission } from '../missions/addMission'
