+++ frontend/src/domain/use_cases/reporting/reopenReporting.ts
@@ -1,6 +1,7 @@
+import { reportingActions } from '@features/Reportings/slice'
+
 import { reportingsAPI } from '../../../api/reportingsAPI'
 import { setToast, ReportingContext } from '../../shared_slices/Global'
-import { reportingActions } from '../../shared_slices/reporting'
 
 import type { Reporting } from '../../entities/reporting'
