+++ frontend/src/features/Reportings/Filters/Table/FilterTags.tsx
@@ -4,10 +4,10 @@ import styled from 'styled-components'
 
 import { ReportingSourceLabels } from '../../../../domain/entities/reporting'
 import { ReportingTargetTypeLabels } from '../../../../domain/entities/targetType'
-import { ReportingsFiltersEnum, reportingsFiltersActions } from '../../../../domain/shared_slices/ReportingsFilters'
 import { useAppDispatch } from '../../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../../hooks/useAppSelector'
 import { useGetControlPlans } from '../../../../hooks/useGetControlPlans'
+import { reportingsFiltersActions, ReportingsFiltersEnum } from '../slice'
 
 export function FilterTags() {
   const { subThemes, themes } = useGetControlPlans()
