+++ frontend/src/features/Reportings/Filters/index.tsx
@@ -4,6 +4,7 @@ import _, { reduce } from 'lodash'
 import { type MutableRefObject, useMemo, useRef } from 'react'
 
 import { MapReportingsFilters } from './Map'
+import { reportingsFiltersActions, ReportingsFiltersEnum } from './slice'
 import { TableReportingsFilters } from './Table'
 import { RTK_DEFAULT_QUERY_OPTIONS } from '../../../api/constants'
 import { useGetControlUnitsQuery } from '../../../api/controlUnitsAPI'
@@ -17,7 +18,6 @@ import {
 } from '../../../domain/entities/reporting'
 import { seaFrontLabels } from '../../../domain/entities/seaFrontType'
 import { ReportingTargetTypeLabels } from '../../../domain/entities/targetType'
-import { reportingsFiltersActions, ReportingsFiltersEnum } from '../../../domain/shared_slices/ReportingsFilters'
 import { useAppDispatch } from '../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../hooks/useAppSelector'
 import { useGetControlPlans } from '../../../hooks/useGetControlPlans'
