+++ frontend/src/features/SideWindow/index.tsx
@@ -5,9 +5,9 @@ import { isDashboardEnabled } from '@features/Dashboard/utils'
 import { REPORTING_EVENT_UNSYNCHRONIZED_PROPERTIES } from '@features/Reportings/components/ReportingForm/constants'
 import { useListenReportingEventUpdates } from '@features/Reportings/components/ReportingForm/hooks/useListenReportingEventUpdates'
 import { ReportingsList } from '@features/Reportings/components/ReportingsList'
+import { reportingActions } from '@features/Reportings/slice'
 import { VigilancesAreasList } from '@features/VigilanceArea/components/VigilanceAreasList'
 import { Icon, NewWindowContext, SideMenu, type NewWindowContextValue } from '@mtes-mct/monitor-ui'
-import { reportingActions } from 'domain/shared_slices/reporting'
 import { omit } from 'lodash'
 import { useEffect, useMemo, useRef, useState, type MutableRefObject } from 'react'
 import { generatePath } from 'react-router'
