+++ frontend/src/store/reducers.ts
@@ -1,6 +1,9 @@
 import { dashboardFiltersPersistedReducer } from '@features/Dashboard/components/DashboardForm/slice'
 import { dashboardReducer } from '@features/Dashboard/slice'
 import { interestPointSlicePersistedReducer } from '@features/InterestPoint/slice'
+import { attachMissionToReportingSliceReducer } from '@features/Reportings/components/ReportingForm/AttachMission/slice'
+import { reportingFiltersPersistedReducer } from '@features/Reportings/Filters/slice'
+import { reportingSliceReducer } from '@features/Reportings/slice'
 import { vigilanceAreaFiltersPersistedReducer } from '@features/VigilanceArea/components/VigilanceAreasList/Filters/slice'
 import { vigilanceAreaPersistedReducer } from '@features/VigilanceArea/slice'
 
@@ -14,8 +17,6 @@ import { mapSliceReducer } from '../domain/shared_slices/Map'
 import { measurementSlicePersistedReducer } from '../domain/shared_slices/Measurement'
 import { missionFiltersPersistedReducer } from '../domain/shared_slices/MissionFilters'
 import { regulatorySlicePersistedReducer } from '../domain/shared_slices/Regulatory'
-import { reportingSliceReducer } from '../domain/shared_slices/reporting'
-import { reportingFiltersPersistedReducer } from '../domain/shared_slices/ReportingsFilters'
 import { semaphoresPersistedReducer } from '../domain/shared_slices/SemaphoresSlice'
 import { administrationTablePersistedReducer } from '../features/Administration/components/AdministrationTable/slice'
 import { backOfficeReducer } from '../features/BackOffice/slice'
@@ -28,7 +29,6 @@ import { mainWindowReducer } from '../features/MainWindow/slice'
 import { attachReportingToMissionsSliceReducer } from '../features/missions/MissionForm/AttachReporting/slice'
 import { missionFormsSliceReducer } from '../features/missions/MissionForm/slice'
 import { missionSliceReducer } from '../features/missions/slice'
-import { attachMissionToReportingSliceReducer } from '../features/Reportings/slice'
 import { sideWindowReducer } from '../features/SideWindow/slice'
 import { stationTablePersistedReducer } from '../features/Station/components/StationTable/slice'
 import { stationReducer } from '../features/Station/slice'
